requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0
matplotlib
pandas
tensorboard
opencv-python
falcon
attrs
waitress
opencv-python
--find-links https://download.openmmlab.com/mmcv/dist/cpu/torch1.9.0/index.html
mmcv-full
mmdet
mmocr
flake8
pytest
pytest-cov
black
pyre-check
ptw